yolo_service/requirements.txt
fastapi[standard]
ultralytics
Pillow
pymongo
firebase-admin
aio-pika
prometheus-client